pyproject.toml
[project]
name = "amp-pool"
version = "0.1.0"
description = "AI Model Pool Service - FastAPI backend"
authors = [
    { name = "zhitao.zhang", email = "zhitao.zhang@waffo.co" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "sqlalchemy>=2.0.0",
    "httpx[socks]>=0.27.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "pymysql>=1.1.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
